REF: `Move nn.bs` to `nn.modules.bs`, Move `Hedger` to `nn.modules` (close #55, close #56)
