+++ pfhedge/__init__.py
@@ -1,1 +1,3 @@
-from .hedger import Hedger
+# Users can import Hedger either as `from pfhedge import Hedger` or
+# `from pfhedge.nn import Hedger`
+from pfhedge.nn import Hedger
