+++ pfhedge/nn/__init__.py
@@ -1,10 +1,11 @@
-from .bs.american_binary import BSAmericanBinaryOption
-from .bs.bs import BlackScholes
-from .bs.european import BSEuropeanOption
-from .bs.european_binary import BSEuropeanBinaryOption
-from .bs.lookback import BSLookbackOption
+from .modules.bs.american_binary import BSAmericanBinaryOption
+from .modules.bs.bs import BlackScholes
+from .modules.bs.european import BSEuropeanOption
+from .modules.bs.european_binary import BSEuropeanBinaryOption
+from .modules.bs.lookback import BSLookbackOption
 from .modules.clamp import Clamp
 from .modules.clamp import LeakyClamp
+from .modules.hedger import Hedger
 from .modules.loss import EntropicLoss
 from .modules.loss import EntropicRiskMeasure
 from .modules.loss import ExpectedShortfall
