+++ pfhedge/nn/modules/bs/lookback.py
@@ -1,23 +1,23 @@
 import torch
 from torch import Tensor
 
-from ..._utils.bisect import bisect
+from pfhedge._utils.bisect import bisect
+
 from ._base import BSModuleMixin
 
 
-class BSAmericanBinaryOption(BSModuleMixin):
-    """Black-Scholes formula for an American Binary Option.
+class BSLookbackOption(BSModuleMixin):
+    """Black-Scholes formula for a lookback option with a fixed strike.
 
     Args:
-        derivative (:class:`pfhedge.instruments.AmericanBinaryOption`, optional):
+        derivative (:class:`pfhedge.instruments.LookbackOption`, optional):
             The derivative to get the Black-Scholes formula.
-        call (bool, default=True): Specifies whether the option is call
-            or put.
+        call (bool, default=True): Specifies whether the option is call or put.
         strike (float, default=1.0): The strike price of the option.
 
     Shape:
         - Input: :math:`(N, *, 4)`, where :math:`*` means any number of additional
-          dimensions. See `features()` for input features.
+          dimensions.  See `features()` for input features.
         - Output: :math:`(N, *, 1)`. Delta of the derivative.
           All but the last dimension are the same shape as the input.
 
@@ -30,8 +30,8 @@ class BSAmericanBinaryOption(BSModuleMixin):
 
         The `forward` method returns delta of the derivative.
 
-        >>> from pfhedge.nn import BSAmericanBinaryOption
-        >>> m = BSAmericanBinaryOption(strike=1.0)
+        >>> from pfhedge.nn import BSLookbackOption
+        >>> m = BSLookbackOption()
         >>> m.features()
         ['log_moneyness', 'max_log_moneyness', 'expiry_time', 'volatility']
         >>> input = torch.tensor([
@@ -39,19 +39,19 @@ class BSAmericanBinaryOption(BSModuleMixin):
         ...     [ 0.00,  0.00, 0.1, 0.2],
         ...     [ 0.01,  0.01, 0.1, 0.2]])
         >>> m(input)
-        tensor([[1.1285],
-                [0.0000],
-                [0.0000]])
+        tensor([[0.9208],
+                [1.0515],
+                [1.0515]])
 
         One can instantiate it using an
-        :class:`pfhedge.instruments.AmericanBinaryOption`.
+        :class:`pfhedge.instruments.LookbackOption`.
 
         >>> from pfhedge.instruments import BrownianStock
-        >>> from pfhedge.instruments import AmericanBinaryOption
-        >>> deriv = AmericanBinaryOption(BrownianStock(), strike=1.1)
-        >>> m = BSAmericanBinaryOption(deriv)
+        >>> from pfhedge.instruments import LookbackOption
+        >>> deriv = LookbackOption(BrownianStock(), strike=1.1)
+        >>> m = BSLookbackOption(deriv)
         >>> m
-        BSAmericanBinaryOption(strike=1.1)
+        BSLookbackOption(strike=1.1)
     """
 
     def __init__(self, derivative=None, call: bool = True, strike: float = 1.0):
@@ -75,20 +75,28 @@ class BSAmericanBinaryOption(BSModuleMixin):
             params.append(f"strike={self.strike}")
         return ", ".join(params)
 
+    def features(self) -> list:
+        return ["log_moneyness", "max_log_moneyness", "expiry_time", "volatility"]
+
+    @torch.enable_grad()
     def delta(
         self,
         log_moneyness: Tensor,
         max_log_moneyness: Tensor,
         expiry_time: Tensor,
         volatility: Tensor,
+        create_graph: bool = False,
     ) -> Tensor:
         """Returns delta of the derivative.
 
         Args:
-            log_moneyness (torch.Tensor): Log moneyness of the underlying asset.
+            log_moneyness (torch.Tensor): Log moneyness of the prices of the underlying
+                asset.
             max_log_moneyness (torch.Tensor): Cumulative maximum of the log moneyness.
             expiry_time (torch.Tensor): Time to expiry of the option.
             volatility (torch.Tensor): Volatility of the underlying asset.
+            create_graph (bool, default=False): If True, graph of the derivative
+                will be constructed. This option is used to compute gamma.
 
         Shape:
             - log_moneyness: :math:`(N, *)`
@@ -97,23 +105,26 @@ class BSAmericanBinaryOption(BSModuleMixin):
             - volatility: :math:`(N, *)`
             - output: :math:`(N, *)`
 
+
         Returns:
-            torch.Tensor
+            Tensor
         """
-        s, m, t, v = map(
-            torch.as_tensor, (log_moneyness, max_log_moneyness, expiry_time, volatility)
-        )
+        prices = self.strike * torch.exp(torch.as_tensor(log_moneyness))
+        prices = Tensor.requires_grad_(prices)
 
-        sqrt2 = torch.sqrt(torch.as_tensor(2.0)).item()
-        d1 = self.d1(s, t, v)
-        d2 = self.d2(s, t, v)
-        c1 = self.N.cdf(d1 / sqrt2)
-        p1 = self.N.pdf(d1 / sqrt2)
-        p2 = self.N.pdf(d2 / sqrt2)
+        s = torch.log(prices / self.strike)
+        m, t, v = map(torch.as_tensor, (max_log_moneyness, expiry_time, volatility))
+        m = torch.max(s, m)  # `s + epsilon` may be greater than `m`
 
-        d = (1 + c1 + (p1 + p2)) / (2 * self.strike)
+        price = self.price(s, m, t, v)
+        delta = torch.autograd.grad(
+            price,
+            inputs=prices,
+            grad_outputs=torch.ones_like(prices),
+            create_graph=create_graph,
+        )[0]
 
-        return torch.where(m < 0, d, torch.zeros_like(d))
+        return delta
 
     @torch.enable_grad()
     def gamma(
@@ -126,10 +137,12 @@ class BSAmericanBinaryOption(BSModuleMixin):
         """Returns gamma of the derivative.
 
         Args:
-            log_moneyness (torch.Tensor): Log moneyness of the underlying asset.
+            log_moneyness (torch.Tensor): Log moneyness of the prices of the
+                underlying asset.
             max_log_moneyness (torch.Tensor): Cumulative maximum of the log moneyness.
             expiry_time (torch.Tensor): Time to expiry of the option.
-            volatility (torch.Tensor): Volatility of the underlying asset.
+            volatility (torch.Tensor):
+                Volatility of the underlying asset.
 
         Shape:
             - log_moneyness: :math:`(N, *)`
@@ -139,7 +152,7 @@ class BSAmericanBinaryOption(BSModuleMixin):
             - output: :math:`(N, *)`
 
         Returns:
-            torch.Tensor
+            Tensor
         """
         prices = self.strike * torch.exp(torch.as_tensor(log_moneyness))
         prices = Tensor.requires_grad_(prices)
@@ -148,7 +161,7 @@ class BSAmericanBinaryOption(BSModuleMixin):
         m, t, v = map(torch.as_tensor, (max_log_moneyness, expiry_time, volatility))
         m = torch.max(s, m)  # `s + epsilon` may be greater than `m`
 
-        delta = self.delta(s, m, t, v)
+        delta = self.delta(s, m, t, v, create_graph=True)
         gamma = torch.autograd.grad(
             delta, prices, grad_outputs=torch.ones_like(prices)
         )[0]
@@ -162,11 +175,10 @@ class BSAmericanBinaryOption(BSModuleMixin):
         expiry_time: Tensor,
         volatility: Tensor,
     ) -> Tensor:
-        """
-        Returns price of the derivative.
+        """Returns price of the derivative.
 
         Args:
-            log_moneyness (torch.Tensor): Log moneyness of the underlying asset.
+            log_moneyness (torch.Tensor): Log moneyness of the prices of the underlying asset.
             max_log_moneyness (torch.Tensor): Cumulative maximum of the log moneyness.
             expiry_time (torch.Tensor): Time to expiry of the option.
             volatility (torch.Tensor): Volatility of the underlying asset.
@@ -185,13 +197,27 @@ class BSAmericanBinaryOption(BSModuleMixin):
             torch.as_tensor, (log_moneyness, max_log_moneyness, expiry_time, volatility)
         )
 
-        sqrt2 = torch.sqrt(torch.as_tensor(2.0)).item()
-        n1 = self.N.cdf(self.d1(s, t, v) / sqrt2)
-        n2 = self.N.cdf(self.d2(s, t, v) / sqrt2)
+        d1 = self.d1(s, t, v)
+        d2 = self.d2(s, t, v)
+        e1 = (s - m + (v ** 2 / 2) * t) / (v * torch.sqrt(t))  # d' in paper
+        e2 = (s - m - (v ** 2 / 2) * t) / (v * torch.sqrt(t))
 
-        p = (1 / 2) * (torch.exp(s) * (1 + n1) + n2)
+        # when max moneyness < strike
+        price_0 = self.strike * (
+            torch.exp(s) * self.N.cdf(d1)
+            - self.N.cdf(d2)
+            + torch.exp(s) * v * torch.sqrt(t) * (d1 * self.N.cdf(d1) + self.N.pdf(d1))
+        )
+        # when max moneyness >= strike
+        price_1 = self.strike * (
+            torch.exp(s) * self.N.cdf(e1)
+            - torch.exp(m) * self.N.cdf(e2)
+            + torch.exp(m)
+            - 1
+            + torch.exp(s) * v * torch.sqrt(t) * (e1 * self.N.cdf(e1) + self.N.pdf(e1))
+        )
 
-        return torch.where(m < 0, p, torch.ones_like(p))
+        return torch.where(m < 0, price_0, price_1)
 
     def implied_volatility(
         self,
@@ -204,22 +230,21 @@ class BSAmericanBinaryOption(BSModuleMixin):
         """Returns implied volatility of the derivative.
 
         Args:
-            log_moneyness (torch.Tensor): Log moneyness of the underlying asset.
+            log_moneyness (torch.Tensor): Log moneyness of the prices of the underlying asset.
             max_log_moneyness (torch.Tensor): Cumulative maximum of the log moneyness.
             expiry_time (torch.Tensor): Time to expiry of the option.
-            volatility (torch.Tensor): Volatility of the underlying asset.
-            precision (float, default=1e-6): Computational precision of the implied
-                volatility.
+            price (torch.Tensor): Price of the derivative.
+            precision (float, default=1e-6): Computational precision of the implied volatility.
 
         Shape:
             - log_moneyness: :math:`(N, *)`
             - max_log_moneyness: :math:`(N, *)`
             - expiry_time: :math:`(N, *)`
-            - volatility: :math:`(N, *)`
+            - price: :math:`(N, *)`
             - output: :math:`(N, *)`
 
         Returns:
-            torch.Tensor
+            Tensor
         """
         s, m, t, p = map(
             torch.as_tensor, (log_moneyness, max_log_moneyness, expiry_time, price)
