+++ tests/nn/bs/test_european_binary.py
@@ -88,9 +88,9 @@ class TestBSEuropeanBinaryOption(_TestBSModule):
         assert torch.allclose(result, expect)
 
     def test_example(self):
-        from pfhedge import Hedger
         from pfhedge.instruments import BrownianStock
         from pfhedge.instruments import EuropeanBinaryOption
+        from pfhedge.nn import Hedger
 
         deriv = EuropeanBinaryOption(BrownianStock())
         model = BSEuropeanBinaryOption(deriv)
